refactor: rename UserPreferencesRepositoryImpl to AppConfigurationsRepositoryImpl and update data source methods
