--- repository/src/main/java/com/baghdad/repository/UserPreferencesRepositoryImpl.kt
@@ -1,35 +1,0 @@
-package com.baghdad.repository
-
-import com.baghdad.domain.repository.UserPreferencesRepository
-import com.baghdad.repository.datasource.local.LocalUserPreferencesDataSource
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.map
-import java.util.Locale
-import javax.inject.Inject
-
-class UserPreferencesRepositoryImpl @Inject constructor(
-    private val localUserPreferencesDataSource: LocalUserPreferencesDataSource
-) : UserPreferencesRepository {
-
-    override suspend fun isAppInDarkTheme(): Flow<Boolean> {
-        return localUserPreferencesDataSource.isAppInDarkTheme()
-    }
-
-    override suspend fun setDarkTheme(enabled: Boolean) {
-        localUserPreferencesDataSource.setDarkTheme(enabled)
-    }
-
-    override suspend fun getAppLanguage(): Flow<String> {
-        return localUserPreferencesDataSource.getAppLanguage().map {
-            if (it == null) {
-                localUserPreferencesDataSource.setAppLanguage(Locale.getDefault().language)
-                Locale.getDefault().language
-            } else it
-        }
-    }
-
-    override suspend fun setAppLanguage(language: String) {
-        localUserPreferencesDataSource.setAppLanguage(language)
-    }
-
-}
